gcp/modules/functions/function_code/requirements.txt
grpcio
google-cloud-firestore==2.19.0
